Add precision category, reorder typing rules
