+++ src/rules/precision/mod.rs
@@ -1,0 +1,2 @@
+pub mod double_precision;
+pub mod kind_suffixes;
