+++ src/rules/typing/mod.rs
@@ -1,4 +1,3 @@
 pub mod implicit_typing;
 pub mod literal_kinds;
-pub mod real_precision;
 pub mod star_kinds;
